Extract code for executing GTests without a JUnit runner
