+++ test-plugins/org.yakindu.sct.generator.c.test/src/org/yakindu/sct/generator/c/gtest/GTestRunner.java
@@ -11,8 +11,6 @@
 
 package org.yakindu.sct.generator.c.gtest;
 
-import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -27,10 +25,7 @@ import java.util.Map.Entry;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import junit.framework.AssertionFailedError;
-
-import org.eclipse.core.resources.IContainer;
-import org.eclipse.core.resources.IResource;
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.FileLocator;
 import org.eclipse.core.runtime.IPath;
@@ -47,6 +42,8 @@ import org.junit.runners.model.InitializationError;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.FrameworkUtil;
 
+import junit.framework.AssertionFailedError;
+
 /**
  * @author Andreas Unger
  * 
@@ -56,7 +53,6 @@ public class GTestRunner extends Runner {
 	private static final Pattern TEST_PATTERN = Pattern.compile("TEST(?:_F)?\\s*\\(\\s*(\\w+)\\s*,\\s*(\\w+)\\s*\\)");
 	private static final Pattern SL_COMMENT_PATTERN = Pattern.compile("//.*(?:\\r?\\n|\\z)");
 	private static final Pattern ML_COMMENT_PATTERN = Pattern.compile("(?:/\\*(?:[^*]|(?:\\*+[^*/]))*\\*+/)|(?://.*)");
-	private static final Pattern TEST_OUTPUT_PATTERN = Pattern.compile("\\[\\s*(\\w+)\\s*\\] (\\w+)\\.(\\w+)");
 
 	private Class<?> testClass;
 
@@ -201,120 +197,23 @@ public class GTestRunner extends Runner {
 		}
 		String targetProject = testClass.getAnnotation(GTest.class).testBundle();
 		IPath programPath = new Path(targetProject).append(program);
-		IResource programFile = ResourcesPlugin.getWorkspace().getRoot().findMember(programPath);
-		IContainer programContainer = programFile.getParent();
-		if (!programContainer.isAccessible()) {
-			throw new RuntimeException(
-					"Test program container " + programContainer.getLocation().toOSString() + " inaccessible");
-		}
-
-		File directory = programContainer.getLocation().toFile();
-		Process process = new ProcessBuilder(programFile.getLocation().toOSString()).redirectErrorStream(true)
-				.directory(directory).start();
-		BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-
-		boolean started = false;
-		boolean running = false;
-		StringBuilder message = new StringBuilder();
-		String line;
-		while ((line = reader.readLine()) != null) {
-			if (line.startsWith("[====")) {
-				if (started) {
-					started = false;
-					// Eat remaining input
-					char[] buffer = new char[4096];
-					while (reader.read(buffer) != -1);
-					break;
-				}
-				started = true;
-			} else {
-				TestOutput testOutput = parseTestOutput(line);
-				if (testOutput != null) {
-					Description description = testOutput.toDescription();
-					switch (testOutput.getStatus()) {
-						case TestOutput.RUN :
-							running = true;
-							message.setLength(0);
-							notifier.fireTestStarted(description);
-							break;
-						case TestOutput.OK :
-							running = false;
-							notifier.fireTestFinished(description);
-							break;
-						default :
-							running = false;
-							notifier.fireTestFailure(
-									new Failure(description, new AssertionFailedError(message.toString())));
-							notifier.fireTestFinished(description);
-							break;
-					}
-				} else if (running) {
-					message.append(line);
-					message.append("\n");
-				}
+		IFile programFile = ResourcesPlugin.getWorkspace().getRoot().getFile(programPath);
+		
+		GTestExecutor executor = new GTestExecutor(testClass) {
+			@Override
+			protected void testStarted(Description desc) {
+				notifier.fireTestStarted(desc);
 			}
-		}
-
-		process.waitFor();
-
-		if (started) {
-			throw new RuntimeException("Test quit unexpectedly (exit status " + process.exitValue() + "):\n" + message);
-		}
-	}
-
-	private TestOutput parseTestOutput(String s) {
-		Matcher matcher = TEST_OUTPUT_PATTERN.matcher(s);
-		if (matcher.find()) {
-			String statusString = matcher.group(1);
-			int status;
-			if ("RUN".equals(statusString)) {
-				status = TestOutput.RUN;
-			} else if ("OK".equals(statusString)) {
-				status = TestOutput.OK;
-			} else {
-				status = TestOutput.FAILED;
+			@Override
+			protected void testFinished(Description desc) {
+				notifier.fireTestFinished(desc);
 			}
-			String testCaseName = matcher.group(2);
-			String testName = matcher.group(3);
-			return new TestOutput(status, testCaseName, testName);
-		}
-		return null;
-	}
-
-	private class TestOutput {
-
-		public static final int RUN = 0;
-		public static final int OK = 1;
-		public static final int FAILED = 2;
-
-		private int status;
-
-		private String testCaseName;
-		private String testName;
-
-		public TestOutput(int status, String testCaseName, String testName) {
-			this.status = status;
-			this.testCaseName = testCaseName;
-			this.testName = testName;
-		}
-
-		/**
-		 * @return the status
-		 */
-		public int getStatus() {
-			return status;
-		}
-
-		public Description toDescription() {
-			String name;
-			if (testCases.size() == 1) {
-				name = testName;
-			} else {
-				name = testCaseName + "." + testName;
+			@Override
+			protected void testFailed(Description desc, String message) {
+				notifier.fireTestFailure(
+						new Failure(desc, new AssertionFailedError(message.toString())));
 			}
-			return Description.createTestDescription(testClass, name);
-		}
-
+		};
+		executor.execute(programFile);
 	}
-
 }
